Starting to incorporate an R package and build on windows again, so fairly radical build changes
-Changed CMAKE_MODULE_PATH
-Added dependency on mpir in windows case
-Added residualConnectivityConfig.cmake, which does different things
depending on windows / linux, unfortunately.
-Fixed OpenMP build flags on windows - You shouldn't add these as a
library on windows, just on other platforms.
-Changed @only to @ONLY in a couple of places
-Changed an OpenMP loop variable to signed
-Fixed a gcc intrinsic in gridCountSpecificSize2
-Only depend on GraphViz if it can be found
-Added missing move constructor
-Changed a couple of float types to double
-Fixed dependency on alloca in transferMatrixCommon

+++ residualConnectivityCommon/subObs/withFinalConditioning.hpp
@@ -22,7 +22,7 @@ namespace discreteGermGrain
 				}
 				const Context& context = object.getContext();
 				double openProbability = context.getOperationalProbabilityD();
-				boost::bernoulli_distribution<float> bern(openProbability);
+				boost::bernoulli_distribution<double> bern(openProbability);
 				//Construct helper graph, containing everything except FIXED_OFF vertices. This is because in order to compute the biconnected components / articulation vertices, we need an actual graph object, not just a state vector
 				subGraphType graph;
 				std::vector<int> graphVertices;
